Rev1

+++ trunk/WebApplication1/src/java/DAO/NixlinksDAO.java
@@ -18,5 +18,5 @@ public interface NixlinksDAO {
 
     public List getAllNixlink(int firstresult, int maxresult) throws SQLException;
 
-    public int getAllNixlinkCount() throws SQLException;
+    public Long getAllNixlinkCount() throws SQLException;
 }
